Moved routes/index.js to router.js
